requirements.txt
flask==2.0.1
python-dotenv==0.19.1
flask-wtf==1.0.0
werkzeug==2.0.1